add some 64-bit file support via --enable-linux-lfs and use doubles in seek()
git-svn-id: https://svn.r-project.org/R/trunk@31119 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/dounzip.c
@@ -308,7 +308,7 @@ static size_t null_write(const void *ptr, size_t size, size_t nitems,
     return 0; /* -Wall */
 }
 
-static long null_seek(Rconnection con, int where, int origin, int rw)
+static long null_seek(Rconnection con, double where, int origin, int rw)
 {
     error("seek not enabled for this connection");
     return 0; /* -Wall */
